Fixed by refactoring FoundationDbGrpcFacade to the lionrock package.
