+++ lionrock-foundationdb-client/src/test/java/io/github/panghy/lionrock/client/foundationdb/impl/AbstractFoundationDBClientTests.java
@@ -4,8 +4,9 @@ import com.apple.foundationdb.Database;
 import com.apple.foundationdb.KeySelector;
 import com.apple.foundationdb.KeyValue;
 import com.apple.foundationdb.Transaction;
+import io.github.panghy.lionrock.FoundationDbGrpcFacade;
 import io.github.panghy.lionrock.client.foundationdb.RemoteFoundationDBDatabaseFactory;
-import io.github.panghy.lionrock.foundationdb.FoundationDbGrpcFacade;
+import io.github.panghy.lionrock.tests.AbstractGrpcTest;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import org.junit.jupiter.api.AfterEach;
