+++ lionrock-foundationdb-server/src/main/java/io/github/panghy/lionrock/Configuration.java
@@ -1,4 +1,4 @@
-package io.github.panghy.lionrock.foundationdb;
+package io.github.panghy.lionrock;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
