+++ lionrock-foundationdb-server/src/test/java/io/github/panghy/lionrock/GrpcTestSuite.java
@@ -1,4 +1,4 @@
-package io.github.panghy.lionrock.foundationdb;
+package io.github.panghy.lionrock;
 
 import org.junit.platform.suite.api.SelectPackages;
 import org.junit.platform.suite.api.Suite;
